:reycle:(depth metrics)! centralize depth metrics in metrics.depth
